+++ src/HappyEngine/code/content/IModelLoader.h
@@ -37,17 +37,17 @@ public:
   
     virtual bool load(const std::string& path, const gfx::BufferLayout& vertLayout, bool allowByteIndices = true) = 0;
 
-    virtual uint32 getNumMeshes() const = 0;
+    virtual size_t getNumMeshes() const = 0;
     virtual const std::string& getMeshName(uint32 mesh) const = 0;
 
     virtual const void* getVertices(uint32 mesh) const = 0;
-    virtual uint32 getNumVertices(uint32 mesh) const = 0;
+    virtual size_t getNumVertices(uint32 mesh) const = 0;
 
     virtual const he::ObjectList<gfx::Bone>& getBones(uint32 mesh) const = 0;
 
     virtual const void* getIndices(uint32 mesh) const = 0;
     virtual gfx::IndexStride getIndexStride(uint32 mesh) const = 0;
-    virtual uint32 getNumIndices(uint32 mesh) const = 0;
+    virtual size_t getNumIndices(uint32 mesh) const = 0;
 };
 
 } } } //end namespace
